refactor(tls): extract tofu tests to sibling file
